morpheus-server/requirements.txt
accelerate==0.15.0
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==22.10.0
boto3==1.26.58
botocore==1.29.58
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
diffusers==0.12.0
exceptiongroup==1.1.0
fastapi==0.86.0
filelock==3.9.0
firebase-admin==6.0.1
flake8==5.0.4
ftfy==6.1.1
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.2.4
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.4
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.1
pytest-faulthandler==2.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.43
starlette==0.20.4
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
trio==0.22.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.6
zipp==3.11.0
